Finish install.sh

+++ install.sh
@@ -4,39 +4,35 @@
 # This scripts installs applications and sets up development environment
 ############################
 
-# shellcheck source=scripts/list-from-file.sh
-source scripts/list-from-file.sh
+# shellcheck source=/dev/null
+
+GREEN='\e[32m'
+NC='\e[0m'
+
+sudo chmod u+x scripts/list-git-release.sh
 
 # Check if curl is installed, if not install it
 type -p curl >/dev/null || sudo apt install curl -y
 
+# Check if xargs is installed, if not install it
+type -p xargs >/dev/null || sudo apt install xargs -y
+
 # Install apt packages
-echo
-echo "Setting Up APT Packages..."
-APT_PACKAGES="$(list_from_file lists/apt-packages.txt)"
-for package in $APT_PACKAGES; do
-	sudo apt install -y "$package"
-done
+echo -e "\n${GREEN}Setting Up APT Packages...${NC}"
+xargs sudo apt install -y <lists/apt-packages.txt
 
 # Install or update nvm
-echo
-echo "Setting Up NVM..."
 sudo chmod u+x ~/dotfiles/scripts/install-nvm.sh
 ./scripts/install-nvm.sh
-
-# get npm packages to install
-NPM_PACKAGES="$(list_from_file lists/npm-packages.txt)"
+source ~/.bashrc
 
 # Install nvm node versions
 for version in lts/* node; do
-	echo
-	echo "Setupging Up Node Version: $version..."
+	echo -e "\n${GREEN}Setting Up Node Version: $version...${NC}"
 	nvm install "$version" &&
-		for package in $NPM_PACKAGES; do
-			sudo apt install -y "$package"
-		done &&
+		xargs npm install -g <lists/npm-packages.txt &&
 		corepack enable &&
-		corepack prepare yarn@latest --activate &&
+		corepack prepare yarn@stable --activate &&
 		corepack prepare pnpm@latest --activate
 done
 nvm alias lts/* default
@@ -45,15 +41,13 @@ source ~/.bashrc
 
 # Install Deno
 if ! command -v deno &>/dev/null; then
-	echo
-	echo "Installing Deno..."
+	echo -e "\n${GREEN}Installing Deno...${NC}"
 	curl -fsSL https://deno.land/install.sh | sh
 fi
 
 # install gh cli
 if ! command -v gh &>/dev/null; then
-	echo
-	echo "Installing GitHub CLI..."
+	echo -e "\n${GREEN}Installing GitHub CLI...${NC}"
 	curl -fsSL https://cli.github.com/packages/githubcli-archive-keyring.gpg | sudo dd of=/usr/share/keyrings/githubcli-archive-keyring.gpg &&
 		sudo chmod go+r /usr/share/keyrings/githubcli-archive-keyring.gpg &&
 		echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/githubcli-archive-keyring.gpg] https://cli.github.com/packages stable main" | sudo tee /etc/apt/sources.list.d/github-cli.list >/dev/null &&
@@ -64,8 +58,7 @@ fi
 
 # install vs code
 if ! command -v code &>/dev/null; then
-	echo
-	echo "Installing VS Code..."
+	echo -e "\n${GREEN}Installing VS Code...${NC}"
 	sudo apt-get install wget gpg &&
 		wget -qO- https://packages.microsoft.com/keys/microsoft.asc | gpg --dearmor >packages.microsoft.gpg &&
 		sudo install -D -o root -g root -m 644 packages.microsoft.gpg /etc/apt/keyrings/packages.microsoft.gpg &&
@@ -78,16 +71,14 @@ fi
 
 #install postman
 if ! command -v postman &>/dev/null; then
-	echo
-	echo "Installing Postman"
+	echo -e "\n${GREEN}Installing Postman...${NC}"
 	curl https://gist.githubusercontent.com/SanderTheDragon/1331397932abaa1d6fbbf63baed5f043/raw/postman-deb.sh | sh &&
 		source ~/.bashrc
 fi
 
 # install insomnia
 if ! command -v insomnia &>/dev/null; then
-	echo
-	echo "Installing Insomnia"
+	echo -e "\n${GREEN}Installing Insomnia...${NC}"
 	wget "https://updates.insomnia.rest/downloads/ubuntu/latest?app=com.insomnia.app&source=website" -O ./insomnia.deb &&
 		sudo apt install ./insomnia.deb &&
 		rm ./insomnia.deb
@@ -95,30 +86,30 @@ fi
 
 # install azure data studio
 if ! command -v azuredatastudio &>/dev/null; then
-	echo
-	echo "Installing Azure Data Studio..."
+	echo -e "\n${GREEN}Installing Azure Data Studio...${NC}"
 	wget https://go.microsoft.com/fwlink/?linkid=2215528 -O ./aszure-data-studio.deb &&
 		sudo apt install ./azure-data-studio.deb &&
 		rm ./azure-data-studio.deb
 fi
 
 # install anydesk
-read -p "Do you want to install AnyDesk (y/n)? " -n 1 -r
-echo
-if [[ $REPLY =~ ^[Yy]$ ]]; then
+if ! command -v anydesk &>/dev/null; then
 	echo
-	echo "Installing AnyDesk..."
-	wget -qO - https://keys.anydesk.com/repos/DEB-GPG-KEY | apt-key add - &&
-		echo "deb http://deb.anydesk.com/ all main" >/etc/apt/sources.list.d/anydesk-stable.list &&
-		sudo apt update &&
-		sudo apt install anydesk
+	read -p "Do you want to install AnyDesk (y/n)? " -n 1 -r
+	echo
+	if [[ $REPLY =~ ^[Yy]$ ]]; then
+		echo -e "\n${GREEN}Installing AnyDesk...${NC}"
+		wget -qO - https://keys.anydesk.com/repos/DEB-GPG-KEY | apt-key add - &&
+			echo "deb http://deb.anydesk.com/ all main" >/etc/apt/sources.list.d/anydesk-stable.list &&
+			sudo apt update &&
+			sudo apt install anydesk
 
+	fi
 fi
 
 # install mongodb compass
 if ! command -v mongodb-compass &>/dev/null; then
-	echo
-	echo "Installing MongoDB Compass..."
+	echo -e "\n${GREEN}Installing MongoDB Compass...${NC}"
 	wget https://downloads.mongodb.com/compass/mongodb-compass_1.34.2_amd64.deb -O compass.deb &&
 		sudo apt install ./compass.deb &&
 		rm ./compass.deb
@@ -129,17 +120,19 @@ fi
 xdg-open https://www.dropbox.com/install?os=lnx
 
 # Setup Themes
+sudo chmod u+x scripts/install-themes.sh
 scripts/install-themes.sh
 
 # Setup Fonts
+sudo chmod u+x scripts/install-fonts.sh
 scripts/install-fonts.sh
 
 # Setup Dotfiles
-echo
-echo "Creating symlinks for dotfiles..."
-cp -rsTv ~/dotfiles/config ~/
+echo -e "\n${GREEN}Creating symlinks for dotfiles...${NC}"
+cp -rsTvf ~/dotfiles/config ~/
 
 # Install Completions
+sudo chmod u+x scripts/install-completions.sh
 scripts/install-completions.sh
 
 source ~/.bashrc
